rename album_art.rs
