+++ src/player/view/cover_art.rs
